requirements.txt
pytube
openai
python-dotenv
opencv-python
moviepy